+++ game/stats/secondary_stats.py
@@ -24,7 +24,7 @@ class SecondaryStats:
   
   def __mul__(self, other):
     if isinstance(other, SecondaryStats):
-      return SecondaryStats(**dict_op(self.__dict__, other.__dict__, lambda a, b: a * b))
-    return SecondaryStats(**{k: v * other for k, v in self.__dict__.items()})
+      return SecondaryStats(**dict_op(self.__dict__, other.__dict__, lambda a, b: int(a * b)))
+    return SecondaryStats(**{k: int(v * other) for k, v in self.__dict__.items()})
 
 SecondaryStats.One = SecondaryStats(1, 1, 1, 1, 1, 1, 1, 1, 1, 1)
